Moved shared mock data to its own file.
