+++ __mocks__/global.ts
@@ -1,0 +1,5 @@
+// Mock Site Name
+export const siteTitle = 'Graeme Pirie';
+
+// Mock Tagline
+export const siteDescription = 'Software Engineer';
