+++ __mocks__/header.ts
@@ -3,56 +3,51 @@ import { Menu } from 'types';
 
 // Mock Navigation Menu
 export const navMenuMock: Menu = {
-        id: "dGVybTo0",
-        name: "Header",
-        menuItems: {
-            nodes: [
-                {
-                    cssClasses: ["home-link"],
-                    id: "cG9zdDoyMg==",
-                    label: "Home",
-                    target: null,
-                    uri: "/",
-                    title: null,
-                },
-                {
-                    cssClasses: [],
-                    id: "cG9zdDoxNA==",
-                    label: "About",
-                    target: null,
-                    uri: "/about/",
-                    title: null,
-                },
-                {
-                    cssClasses: [],
-                    id: "cG9zdDoyMQ==",
-                    label: "Experience",
-                    target: null,
-                    uri: "/experience/",
-                    title: null,
-                },
-                {
-                    cssClasses: ["linkedin"],
-                    id: "cG9zdDoxNg==",
-                    label: "Linked In",
-                    target: "_blank",
-                    uri: "https://www.linkedin.com/in/graeme-pirie-08625275/",
-                    title: null,
-                },
-                {
-                    cssClasses: ["github"],
-                    id: "cG9zdDoxOA==",
-                    label: "GitHub",
-                    target: "_blank",
-                    uri: "https://github.com/gpirie",
-                    title: null,
-                },
-            ],
-        },
-    };
+    id: "dGVybTo0",
+    name: "Header",
+    menuItems: {
+        nodes: [
+            {
+                cssClasses: ["home-link"],
+                id: "cG9zdDoyMg==",
+                label: "Home",
+                target: null,
+                uri: "/",
+                title: null,
+            },
+            {
+                cssClasses: [],
+                id: "cG9zdDoxNA==",
+                label: "About",
+                target: null,
+                uri: "/about/",
+                title: null,
+            },
+            {
+                cssClasses: [],
+                id: "cG9zdDoyMQ==",
+                label: "Experience",
+                target: null,
+                uri: "/experience/",
+                title: null,
+            },
+            {
+                cssClasses: ["linkedin"],
+                id: "cG9zdDoxNg==",
+                label: "Linked In",
+                target: "_blank",
+                uri: "https://www.linkedin.com/in/graeme-pirie-08625275/",
+                title: null,
+            },
+            {
+                cssClasses: ["github"],
+                id: "cG9zdDoxOA==",
+                label: "GitHub",
+                target: "_blank",
+                uri: "https://github.com/gpirie",
+                title: null,
+            },
+        ],
+    },
+};
 
-// Mock Site Name
-export const siteTitle = 'Graeme Pirie';
-
-// Mock Tagline
-export const siteDescription = 'Software Engineer';
